requirements.txt
tqdm
ultralytics
pyyaml
streamlit
opencv-python
pytesseract
